extract module: fixtures::init
